# src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  title: BlueBubbles ChatGPT Agent OpenAPI spec
  description: Agent that bridges ChatGPT and iMessage
  version: 0.0.1
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/subscription/get.subscription:
    get:
      operationId: subscriptionGet
      tags:
        - subscription
      summary: Gets current subscription status
      description: Returns billing and entitlement state for the authenticated website account.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current subscription status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionSummaryResponse'
        '401':
          description: Authentication required
  /api/v1/subscription/createCheckout.subscriptionCheckouts:
    post:
      operationId: subscriptionCreateCheckout
      tags:
        - subscription
      summary: Creates a subscription checkout session
      description: Creates a provider checkout session for the authenticated website account.
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriptionCheckoutRequest'
      responses:
        '200':
          description: Checkout created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionCheckoutResponse'
        '400':
          description: Invalid checkout request
        '401':
          description: Authentication required
  /api/v1/subscription/createPortal.subscriptionPortals:
    post:
      operationId: subscriptionCreatePortal
      tags:
        - subscription
      summary: Creates a subscription management portal session
      description: Creates a provider portal session for the authenticated website account.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Portal session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionPortalResponse'
        '401':
          description: Authentication required
        '404':
          description: No subscription found
  /api/v1/subscription/receiveWebhook.subscriptionProviderEvents:
    post:
      operationId: subscriptionReceiveWebhook
      tags:
        - subscription
      summary: Receives payment provider webhook events
      description: Receives raw provider webhook payloads. The provider adapter validates signatures before processing.
      parameters:
        - name: provider
          in: query
          description: Subscription provider key, such as btcpay.
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Webhook processed, ignored, or duplicate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionProviderWebhookResponse'
        '401':
          description: Invalid provider webhook signature
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionProviderWebhookResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SubscriptionPlan:
      type: object
      required:
        - key
        - display_name
        - price_amount
        - currency
        - billing_interval
        - provider
        - active
      properties:
        key:
          type: string
        display_name:
          type: string
        description:
          type: string
          nullable: true
        price_amount:
          type: string
          description: Decimal plan price as configured by the app.
        currency:
          type: string
        billing_interval:
          type: string
        provider:
          type: string
        active:
          type: boolean
...
    SubscriptionCheckoutRequest:
      type: object
      properties:
        plan_key:
          type: string
          nullable: true
        provider:
          type: string
          nullable: true
    SubscriptionCheckoutResponse:
      type: object
      required:
        - checkout_session_id
        - provider
        - plan_key
        - checkout_url
      properties:
        checkout_session_id:
          type: string
        provider:
          type: string
        plan_key:
          type: string
        checkout_url:
          type: string
          format: uri-reference
        expires_at:
          type: string
          format: date-time
          nullable: true
    SubscriptionPortalResponse:
      type: object
      required:
        - provider
        - portal_url
      properties:
        provider:
          type: string
        portal_url:
          type: string
          format: uri-reference
    SubscriptionSummaryResponse:
      type: object
      required:
        - account_id
        - is_premium
        - entitlement_source
        - plans
        - subscriptions
      properties:
        account_id:
          type: string
        is_premium:
          type: boolean
        entitlement_source:
          type: string
        premium_until:
          type: string
          format: date-time
          nullable: true
        plans:
          type: array
          items:
            $ref: '#/components/schemas/SubscriptionPlan'
        subscriptions:
          type: array
          items:
            $ref: '#/components/schemas/AdminSubscriptionItem'
        latest_checkout:
          type: string
          nullable: true
    SubscriptionProviderWebhookResponse:
      type: object
      required:
        - status
        - message
      properties:
        status:
          type: string
        message:
          type: string
          nullable: true
        account_id:
          type: string
          nullable: true
        subscription_id:
          type: string
          nullable: true
...
    AdminSubscriptionItem:
      type: object
      required:
        - subscription_id
        - account_id
        - provider
        - plan_key
        - status
        - updated_at
      properties:
        subscription_id:
          type: string
        account_id:
          type: string
        account_bucket:
          type: string
          nullable: true
        provider:
          type: string
        plan_key:
          type: string
        status:
          type: string
        provider_status:
          type: string
          nullable: true
        provider_subscription_id:
          type: string
          nullable: true
        provider_customer_id:
          type: string
          nullable: true
        current_period_end:
          type: string
          format: date-time
          nullable: true
        grace_period_end:
          type: string
          format: date-time
          nullable: true
        management_url:
          type: string
          nullable: true
        last_synced_at:
          type: string
          format: date-time
          nullable: true
        updated_at:
          type: string
          format: date-time
...